Update patch set 6
Patch Set 6:

Is there a way that we can make this more generic? Could this be the "ServicesCore" domain. This is AOSP so putting Google specifics in here seems odd. For example, microg[1] "ServicesCore" would likely run in this domain on non-gms devices. Is there a flag in packagemanager that we can use to send gmscore or microg into this new domain?

[1] https://microg.org/

Patch-set: 6
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>