Update patch set 8
Patch Set 8:

(1 comment)

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > Is there a way that we can make this more generic? Could this be the "ServicesCore" domain. This is AOSP so putting Google specifics in here seems odd. For example, microg[1] "ServicesCore" would likely run in this domain on non-gms devices. Is there a flag in packagemanager that we can use to send gmscore or microg into this new domain?
> > 
> > [1] https://microg.org/
> 
> microg would also need its own line in seapp_contexts, routing this into the gmscore domain.
> 

Calling it the gmscore domain is confusing if we expect other apps to run here on non-gms devices.

> To be clear, this isn't a dependency from AOSP to Google services. This just says, if the Google service exists, then run it in a particular sandbox. It doesn't require the Google service to exist. I would prefer that we not introduce any new flags to handle this, and instead based it off the current package name system.

This change is creating neverallow exceptions for gmscore, which means that another app which is intended to do similar things on non-GMS devices would fail CTS or be required to run in the gmscore domain. We should either generalize this, or remove the neverallow tests.

Patch-set: 8
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>