Update patch set 2
Patch Set 2:

The pattern of this change follows an earlier change for dalvik.vm.dex2oat-*threads*. Those properties are now listed in system/sepolicy/prebuilts. Is it safe to assume these get revised when the API level changes and no action is needed here to deal with this?

Patch-set: 2