Update patch set 1
Patch Set 1:

I thought zygote was the one who was suppose to be mounting on /data/media, so that zygote could determine the visibility of files at that location. See, for example,

https://android-review.googlesource.com/c/platform/system/sepolicy/+/891993
and
b/123533205

It feels like an error prone situation to have both vold and zygote mounting onto this location, no? Architecturally, who is suppose to be responsible for handling these files?

Patch-set: 1
CC: Gerrit User 1101746 <1101746@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1013508 <1013508@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>