Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I thought zygote was the one who was suppose to be mounting on /data/media, so that zygote could determine the visibility of files at that location. See, for example,
> 
> https://android-review.googlesource.com/c/platform/system/sepolicy/+/891993
> and
> b/123533205
> 
> It feels like an error prone situation to have both vold and zygote mounting onto this location, no? Architecturally, who is suppose to be responsible for handling these files?

Sorry I wasn't clear enough; the location is not "/data/media", but rather a subdirectory under it; eg "/data/media/Android/obb" for OBBs. The reason to do this in vold and not in Zygote is that this means we only have to do this once (in the root namespace that Zygote and other apps spawn from), instead of doing the same bind mount on every mount.

That said, vold is the one currently mounting on top of "/data/media" as well, and I'm not sure why Zygote should ever have to - it looked like a bug in sdcardfs that was never root caused? Either way, looks the CL you pointed at is planned to be reverted as part of b/130812417 - I'm happy to pick that up.

Patch-set: 1