Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> Tomasz, could you
> 
> 1) define a new property context instead of reusing exported_*. We are going to remove those property contexts.
> 2) make it a public system property (using system_public_prop macro) in system/sepolicy/public/property.te
> 3) allow vendor_init to write the property context, in system/sepolicy/public/vendor_init.te
> 4) neverallow domain -init -vendor_init to write the property
> 5) allow the car service app to read it

Done.

> Patch Set 1:
> 
> why not define in-band in the HAL?

I've considered that, but a read-only property seems like a better solution for this specific use case. For development purposes this needs to be changed without rebuilding VHAL binaries (switching between mock HAL and real HAL) but building a whole manager HAL just for switching at runtime would be an overkill never used in a production vehicle. On top of that, we would still need multiple instances of the HAL that would be used in some production vehicles (but not in a runtime-switching manner).

Patch-set: 3