Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Added permissions in system_server.te, not sure if this is sufficient, please review the change, thanks!
> > 
> > Have you tested it? "Test: build" isn't really sufficient here. What tests were run? Did you verify that the directory exists, has the proper permissions, and that configupdater is allowed to write there?
> 
> And that no new/related selinux denials occur?

I wanted to test but I don't have a chance to let ConfigUpdater to write it due to some blocking bugs in ConfigUpdater. Our project is due, and I would be appreciated if some missing configuration could be figured out in the code review. If not, I understand and I will perform the tests right after the bugs are fixed.

Patch-set: 4
Reviewer: Gerrit User 1131623 <1131623@85c56323-6fa9-3386-8a01-6480fb634889>