Update patch set 1
Patch Set 1:

> Can't a generic domain be defined? I'm pretty sure this rule must be allowed for every updater app that supports A/B.
> Then gmscore can easily extend it in its type definition ( like any other domain )

There are two ways that you could assign a set of rules to multiple different SELinux domains:

1) Use SELinux attributes, and assign the types to the specific attributes.
2) Emulate attributes using m4 macros.

> 
> You also have an example update app in aosp, this will likely break it out of the box.

We'll need to make sure this app is also in it's own domain, at some future point.

> 
> It would be optimal to have something like this imo.
> Other than that I haven't seen updater apps diverge a lot from standard guidelines, our own for example is using all your APIs.
> 
> I see that it'll be hard for you to handle every scenario, but a bare minimum set of rules will be required if the plan is to clean up the bloated priv_app domain.

We don't want to introduce more abstraction (and thus more complexity) when we can keep things simple. Priv-apps are expected to integrate deeply with the OS, and as such, application developers creating priv-apps are expected to be able to adjust the rules associated with their apps. By analogy, this is exactly what we require for daemons started from init.

If you think there are ways that you can provide an abstraction without making SELinux policy overly complex, we're happy to review such a change.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>