Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Looks like we have some plan to clean up SELinux contexts.
> > 
> > https://docs.google.com/document/d/1wwPe88CLetEgYdupIq3JwV6FHd_WL7g2n9X9dSfwV-4/edit#heading=h.ta91qzd449tr  (internal doc)
> > 
> > +Inseob to ensure the current change is ok for now.
> 
> Adding more exported_default_prop doesn't seem good. Although it hasn't been removed for now, how about introducing name without terms like "exported" or "default"?

I think we'd better keep this as is for now because (1) the added properties belong to the group of existing ro.[vendor|odm].build.* properties and (2) we need to port the change back to android10-gsi.
If we want to re-organize existing properties, we should plan it carefully and make sure we don't break old devices. WDYT?

Patch-set: 2
Reviewer: Gerrit User 1061057 <1061057@85c56323-6fa9-3386-8a01-6480fb634889>