Update patch set 1
Patch Set 1: Code-Review-2

This is incorrect. Init already has permission to create this symlink. The problem is the device-specific init.rc file, which has the following lines:

    # Support legacy paths
    symlink /sdcard /mnt/sdcard
    symlink /sdcard /storage/sdcard0

These lines should be deleted from the device specific *.rc file.

What is occurring here is that init is attempting to create this symlink twice, and the second creation is on a different filesystem.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2