Update patch set 1
Patch Set 1: -Code-Review

> Can we not do this? We've been trying to remove these properties since O, and we're finally getting close (all app access has been removed). b/33308258

Oh, I was unaware of this (and thanks to android-blintz for helpfully closing another bug just because it was old or whatever.

So, which mechanism would you recommend we use to give vendors a knob for affecting the internal behavior of the DnsResolver without touching the code we ship within the signed DnsResolver module?

Patch-set: 1
Reviewer: Gerrit User 1273909 <1273909@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Presubmit-Ready=0