access_vectors: Remove entrypoint and execute_no_trans from chr_file
entrypoint and execute_no_trans permissions are only ever checked by the
kernel to regular files.  They were added to the chr_file class when
execmod was added (which can be checked on chr_file) to ensure that it
was assigned the same value as for the file class, since the kernel code
always checked FILE__EXECMOD. However, the policy definitions are not
necessary since the kernel and policy values have been decoupled ever
since dynamic class/perm support was introduced and further with the
move of execmod to the common definitions, they were not even needed
in the kernel.

NB This may have compatibility implications for Android given that
execute_no_trans was being used in neverallow rules on gpu_device:chr_file
for webview_zygote and app_zygote (fixed by this change).
Further, there are neverallow and dontaudit rules on chr_file in cil modules
under prebuilts/api that include both entrypoint and execute_no_trans
permissions, possibly expanded from .te rules using tilde (~) as part
of the conversion to cil. I did not modify those rules in this change,
but they will likely fail to compile without these definitions particularly
since cil has tightened its validity checking of permissions.
Consequently, you may need to defer or never apply this change for
Android for compatibility reasons, which should be harmless.

These were removed from the kernel's classmap by
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=b424485abe2b16580a178b469917a7b6ee0c152a
circa v2.6.36.

The corresponding change was made to refpolicy in
https://github.com/SELinuxProject/refpolicy/commit/8486b8aa83afa7abd94c9338e8845c2cbeb67f31

Test: policy still builds

Change-Id: Ia7117374432f8dfe8d67e89b71b2edac8db3a094
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>