Update patch set 2
Patch Set 2: Code-Review+2 Presubmit-Ready+1

> How does this differ from the previous attempt?
Untrusted apps are still allowed to bind in this one vs https://r.android.com/1212285.

(Not sure how to make that easier to figure out in Gerrit. Maybe this could be a revert of the revert, with the deltas as a patchset 2?)

Patch-set: 2
Reviewer: Gerrit User 1060831 <1060831@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Presubmit-Ready=+1