Update patch set 1
Patch Set 1:

Trying to get bpf config to be correct from the initial bring up.
(currently we rely on stuff getting chowned later on)

Is there an alternative way to grant bpfloader privs so it can change the gid of some pinned bpf maps? Perhaps we could simply make it a member of the groups?  (but I'm not sure how)

Patch-set: 1
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0