Update patch set 1
Patch Set 1:

I don't understand the details of this CL but I suspect that it probably broke this unit test:

https://android.googlesource.com/platform/libcore/+/master/luni/src/test/java/libcore/android/system/OsTest.java#578

That test was introduced in https://r.android.com/135491

Any idea how to update that unit test such that it passes again and still is meaningful?

I'll comment on http://b/148621612 and assign it to you for now.

Patch-set: 1
CC: Gerrit User 1103203 <1103203@85c56323-6fa9-3386-8a01-6480fb634889>