Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> 
> Should this be limited to debuggable builds (userdebug or eng), or all builds? Generally speaking, persisting logging data on production builds is discouraged.

Thanks for the review. This is submitted so that we get a better view of what's going on in droidfood. I'll follow up with another CL to make it only happen with userdebug or eng.

Although, these logs are just as important as the update_engine logs, and I am just following what they do there...

Patch-set: 5