property_contexts: Add interaction helper property label
The new label is used for ro.vendor.cts_interaction_helper_packages,
which is expected to be set by a vendor and then read by a
DeviceInteractionHelperRule from arbitrary CTS tests.  Other similar
properties may be added in the future for other xTS suites, and they
should be able to share the same label.

Bug: 132747867
Test: atest CtsPrintTestCases with r.android.com/1185105
Change-Id: I109c5627afaebee3b2ecbf83b7ed8af9ad747624