Update patch set 10
Patch Set 10:

(2 comments)

> Patch Set 9:
> 
> (1 comment)
> 
> What's the purpose of this change? Why do we need HIDL and AIDL based versions of the same service?

The AIDL version is the "1.1" version of the HIDL HAL, I was originally going to make a 1.1 HIDL revision but was asked to move it to a stable AIDL HAL since it was a very small HAL and that was the direction Android wanted to move the HALs in.

The issue with the 1.0 version is that it is quite hard to make it work with Perfetto/Traceur. The Atrace HAL defines custom vendor tracing categories (corresponding to custom kernel trace points) which users (Perfetto, cmds/atrace) can enable/disable. This works fine for the cmds/atrace commandline tool but for Perfetto (which multiplexes many simultaneous traces onto a single kernel trace) we need to know which custom categories correspond to which custom events in order to route the events to the correct (perfetto) trace. The new "1.1" version fixes this issue. It is intended that vendors need to implement at most one of the two versions.

Patch-set: 10
Reviewer: Gerrit User 1047376 <1047376@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0