Update patch set 4
Patch Set 4: Code-Review+2

we have discussed over the chat. the plan is to piggy-backing the existing property context for now. we will create appropriate property contexts for each of these later when we are about to remove the exported_* props.

Patch-set: 4
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2