Update patch set 1
Patch Set 1:

Joel, this CL was already merged to QPR3 in ag/10777722, which you approved. We need another approval here.

Our OEM partners are having problems with failing CTS-on-GSI tests because we're missing this change, so we're cherry picking it to the android10-gsi branch in AOSP, to fix the failing tests asap.

See b/143146884 and b/153921768 for context.

Patch-set: 1
Reviewer: Gerrit User 1208973 <1208973@85c56323-6fa9-3386-8a01-6480fb634889>