Update patch set 1
Patch Set 1:

We could use /proc/config.gz to detect that it's *not* supported.
(by virtue of things missing)

But we cannot use it to detect that it *is* supported,
because there are kernel patches the lack of which we cannot detect.
(and the lack of which potentially manifests as occasional traffic blackholing)

I think this all becomes moot in S, because we roll it back and we'll simply require devices to support bpf or bpfloader and netd will crash during startup (this means no 4.4 kernel support though, which means no pixel 2 support, but I think that's getting dropped around Christmas 2020 anyway [judging based on when pixel 1 support was dropped last year]).

Patch-set: 1
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0