Update patch set 4
Patch Set 4:

> I really want to remove those exported_* property context which actually is a dumping ground for unrelated system properties.
> 
> I expect something like this
> 
> ro.kernel.ebpf.supported u:object_r:kernel_config_prop:s0 exact bool
> 
> system_vendor_config_prop(kernel_config_prop)  in property.te
> 
> vendor

But this is entirely equivalent (in terms of access patterns) to a ton of other properties that already exist, and right now seems like wrong moment to be changing sepolicy for a single property - it should be done on masse.  Anybody should be able to read it.  And writes need to happen before bpfloader and netd run, which I think means it can't be written on the device at all and instead needs to be built into the image like build fingerprints, and like it it is known at build time (it's a function of the kernel binary/modules).

Patch-set: 4
Reviewer: Gerrit User 1003247 <1003247@85c56323-6fa9-3386-8a01-6480fb634889>