Make genfs/path restrictions symmetrical.
FIXME: in order to submit this, the error messages should be

Before, we had a restriction that /system/* should be system_file_type.
However, this did not restrict things outside of /system/ from being
labeled as system_file_type. So for instance, /product/ could have been
labeled as system_file_type and we did not realize it.

Bug: 154851797
Test: wip
Change-Id: Iff844302382c836f3ce8bdaffe791cd5e7b84d67