Update patch set 4
Patch Set 4: Code-Review-1

It seems that the bootloader implementation was backported to Q/P. IMO, this should correspond to a cuttlefish-specific sepolicy change in the branch that it is built, not a change here. This is what it would be for any other device manufacturer. One will be needed to sepolicy in whatever branch cuttlefish is built anyway, if that branch is to avoid denials for a cf built entirely there.

Patch-set: 4
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1