Update patch set 3
Patch Set 3:

Can you please double check that /sys/kernel/tracing isn't being mounted twice, and the second mount is generating this SELinux denial? The label debugfs_tracing_debug only exists on the debugfs and tracefs filesystems, but the mount point is on the sysfs filesystem, which has a different label.

Usually when these types of denials occur, it's because a double mount is occurring, eg:

mount /sys
mount /sys/kernel/tracing
mount /sys/kernel/tracing

and the second mount is trying to mount on top of a tracefs filesystem, not a sysfs filesystem.

Patch-set: 3
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>