Update patch set 3
Patch Set 3:

> I think Nick's right. On cuttlefish we are mounting it in our vendor init.rc. Can you add a patch to this topic which removes that unnecessary mount?

If we confirm that the double mount is occurring, then the current change, as written, is not necessary.

However, it would be nice if we could add an SELinux "neverallow" compile time assertion, to detect and prevent other similar future changes. Could we add the following to domain.te:

  # the tracefs filesystem is a leaf filesystem, and nothing should
  # ever be mounted on top of it. In particular, to avoid double mounts,
  # a tracefs filesystem should never be mounted on top of a tracefs
  # filesystem.
  neverallow domain debugfs_tracing_debug:dir mounton;

This will help educate others about this problem, and avoid potential regressions.

Patch-set: 3
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>