Update patch set 3
Patch Set 3: Code-Review+2

Thanks nnk@. I think this particular CL is fine for what it is trying to do. i.e. make sure some platform processes can access /vendor/overlay/ for RROs. The drmserver, mediaserver seem to be new found exceptions that need to be added to the list of processes that can access /vendor/overlay.

I have a bigger concern though -

the label 'vendor_overlay_file' was meant to be for the overlays on "vendor side" of the treble boundary. Looking at - 

https://cs.corp.google.com/android/system/sepolicy/private/file_contexts?q=f:system/sepolicy+vendor_overlay_file&sq=package:%5Eandroid$&l=371&dr=C

It seems that the label is now used for /system_ext/overlay, /product/overlay directories that are not vendor partitions. The split has been {system,system_ext,product,oem} <-> {vendor, odm}.

If this has been working, I guess we are being lucky, but we are labeling things on the 'system side' with labels from the 'vendor side'.

When we do it right, may be we find some other coredomains that were granted the exemption, don't really need the exemption.

[I think the problem is related but probably should be fixed separately anyway, so +2ed this one]

Patch-set: 3
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2