Allow access to the "refresh_rate_switching" system property
See ag/11252856 for details. Cherry-pick of aosp/1110734.

Bug: 154892895

Test: Apply the display mode patches described in b/150857674, and use
the screen resolution APK (cl/308216014) to switch display modes. It is
expected that the device will reboot and then the display mode will
change. The scale of the screen will be off. Further work needs to be
done by partner's implementation of switching display modes.

Test: Ran automated tests: RefreshRateConfigsTest, RefreshRateStatsTest,
SchedulerTest.

Exempt-From-Owner-Approval: Backport to TV branch
Change-Id: I8a9c6146c70ac525a1205e966584d7bf20b74b5f