Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> According to b/155836352, the same happens in between system_suspend and dumpstate.
> 
> So, shouldn't we have something like
> 
> allow domain dumpstate:fd use;
> 
> or 
> 
> allow { some_domains ... } dumpstate:fd use;
> 
> in dumpstate.te?

The policy is supposed to be organized so that A.te contains "allow A ....". If we want this, I think we should have an attribute/macro.

On another level, it goes back to making the sepolicy readable/flat as possible/explicit. If we create a lot of macros/attributes, then it obscures what is actually going on. In sepolicy, the important thing is not "we've got the permission to do high-level thing X and Y", it actually matters what is going on underneath. fwiw, I miss how strongly +nnk used to push things in this direction, and that philosophy is why I didn't complain about duplication here.

macros also give a false sense of the system wrt treble mapping files: b/36623345

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1120458 <1120458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0