Update patch set 2
Patch Set 2:

> The policy is supposed to be organized so that A.te contains "allow A ....". If we want this, I think we should have an attribute/macro.
> 

I see that some rules are in the *.te file for the target context, not the source context, especially when the rules are for the same target context from multiple source contexts.

Anyway, I think we should follow the style guideline from the owners. I just don't know what the guideline is. So, over to nnk@ or jeffv@.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>