Update patch set 3
Patch Set 3:

system_ext is an extension of system. It should be generic, i.e. agnostic to device or product, but could be specific to a company that owns the system partition.

That said, we don't want this to be controlled by BOARD_* variables. 

It would be nicer if system_ext sepolicy is from a fixed place like /system/sepolicy/ext/[public|private] or /system/sepolicy/[public|private]/ext.

Patch-set: 3
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>