Update patch set 3
Patch Set 3:

> Or are you insinuating that system_ext should always belong to Google and as such only aosp sepolicy in system/sepolicy should be valid?

No. That's not my intent. system_ext.img is NOT exclusive to Google. Likewise, system.img is also not exclusive to Google; Android is open source :).  My intent was that system and system_ext belong to the entity that owns the fork of Android and should be agnostic to device/hardware.

BOARD_ and PRODUCT_ variables could be set differently per device/hardware. So I think being able to set the sepolicy for system_ext either via BOARD_ or PRODUCT_ gives a false sense that system_ext could be specific to device/hardware.

So, yeah, you are right; that might be achieved by another CL, but that CL will effectively revert this change.

Patch-set: 3
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>