Update patch set 2
Patch Set 2:

As we went through the list of properties to add to SystemProporties,  We decided to not support wake_on_hotplug.  The code that uses it has been removed.
We can leave the property as "unused" but it has no effect.

Patch-set: 2
Reviewer: Gerrit User 1060286 <1060286@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0