Update patch set 3
Patch Set 3: Code-Review-1

Actually I think we should remove the "ce_available" property entirely, since it's no longer relevant in Android R with some of the changes being made to storage.  (It appears to only be used by mkdirs(), which we should instead be guarding based on the mounted state of the targeted storage device.)

Can you leave the property defined as-is in this CL and file a separate bug to track the removal?  (Note that my CR-1 is only related to the storage property; I have no opinion about the other properties.)

Patch-set: 3
Reviewer: Gerrit User 1013508 <1013508@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1