Update patch set 4
Abandoned

Given that we're past SEPolicy freeze for Android 11 and we haven't had any specific requests for this, I'm going to abandon this for now. If we hear from vendors that not being able to access from vendor init is a problem, we can revisit this at that time. Thanks for all your help, reviewers, and sorry for the churn.

Patch-set: 4
Status: abandoned
Tag: autogenerated:gerrit:abandon