Update patch set 2
Patch Set 2:

(3 comments)

We have a problem, bootstat logic appears to have a voting problem between reboot,quescent and cold,powerkey and reboot,software_master.

The bootloader is using _old_ unsupported non-compliant boot reasons which undoubtably also contributes to the problems, but a deep dive will be required to find out why the bootstat logic is not firing as expected.

Patch-set: 2