Update patch set 1
Patch Set 1:

I think this is too broad, especially when the prop access is required by a CTS test.

I think the test should be amended to not read the prop directly. For example, we could let packagemanagerservice to emit the value of the prop via dumpsys and let the cts test to have that info via dumpsys.

Patch-set: 1