Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> What is dlkm? It might be nice to have that explained somewhere. Preferably in a comment within policy so that future-us will know what it is. Better yet, can you name this something that is self-explanatory?

It means dynamically-loadable kernel module. 

BTW, I wonder if we need a new file_contexts files for this. If the path /vendor_dlkm/* is a known path, can we just label them in system/sepolicy/private/file_contexts? (as we already do for most of the vendor/* paths)

Patch-set: 3