Update patch set 3
Patch Set 3:

> But, if this structure changes in the future, wouldn't it be nice that updating itself alone is sufficient?

I think you we'd better not to change the structure in the future. And I think that's feasible given that we have kept the same for other directories under /vendor.

Changing the structure in the future might be challenging even when the file_contexts file is within the partition. You are not sure if the paths are not used outside of the partition.

Patch-set: 3