WIP: HACK: sepolicy: Add dir search and rw permissions on ion_device (for dmabuf heaps)
This is an ugly hack, suggestions for a proper fix would be welcome!

This adds dir search and rw permissions to the ion_device as
with dmabuf heaps we need to be able to check the parent dir
and also open the chardevs rw for ioctls.

We probably need a new heap_device spolicy entry for this?

Change-Id: I945fac2116aaef24490a54bdde861d2c6125e038
Signed-off-by: John Stultz <john.stultz@linaro.org>