Change ro.arch's context to build_prop
ro.arch becomes build_prop, to remove bad context exported2_default_prop.

According to external/cpuinfo/src/arm/linux/chipset.c, it's only used in
exynos devices. It could be removed from AOSP because it's SoC specific.
But it would be better to leave it as long as external/cpuinfo dirctly
refers to it.

Bug: 155844385
Test: m selinux_policy
Change-Id: I0026858193f1c10835bf1e27dc2e356e80f251a3