Introduce app_data_file_type attribute.
This gives us an easy way to refer to all existing or future types
used for app private data files in type= assignments in
seapp_contexts.

Apply the label to all the existing types, then refactor rules to use
the new attribute.

This is intended as a pure refactoring, except that some neverallow
rules are extended to cover types they previous omitted.

Test: presubmits
Bug: 171795911
Change-Id: Iab6018af449dab3b407824e635dc62e3d81e07c9