Survey root file system paths in sepolicy.
The idea of this is to require someone to add something to this list in
order to add a new file system type, but it's quite involved, and some
devices add more file locations.

This CL currently fails with this error on cuttlefish:
  The following types on / must be associated with the "no_fs_type"
  attribute: rootfs
   corresponding to files: /

However, the number of problems introduced when system_ext/product/odm
were introduced were minimal/non-existant, so I will abandon this CL,
but uploading for possible interest.

Bug: 154851797
Test: build (fails)
Change-Id: I14d363be7713837e30c1019cbd2226bc94128012